- Added script to automatize the parametric perf test on clusters - Loader (dff_run) has now the ability to force the distributed protocol: MPI and TCP - Added distributed printer to tag the output of stdout when multiple processes print on screen. To use this feature ff::cout << something << ff::endl; - Loader (dff_run) has now the ability to build an host file for mpi based on the directives described in the dff configuration file. Now the execution of MPI application can be triggered correctly directly from the loader.

+++ ff/distributed/ff_dgroup.hpp
@@ -354,7 +354,6 @@ public:
 
     int run(ff_node* baseBB, bool skip_init=false) override {
         buildFarm(reinterpret_cast<ff_pipeline*>(baseBB));
-
         return ff_farm::run(skip_init);
     }
 
@@ -758,7 +757,7 @@ bool MySet<T>::check_inout(ff_node* node){
     }
 
 void dGroups::consolidateGroups(){
-    for(int i = 0; i < parsedGroups.size(); i++){
+    for(size_t i = 0; i < parsedGroups.size(); i++){
         const G & g = parsedGroups[i];
         if (groups.find(g.name) != groups.end())
             switch (this->usedProtocol){
